apiproject/requirements.txt
asgiref==3.5.2
backports.zoneinfo;python_version<"3.9"
Django==4.0.6
djangorestframework==3.13.1
gunicorn==20.1.0
install==1.3.5
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
whitenoise==6.2.0
